+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [3.2.0] - 2023-08-23
+- Upgraded python-lib to v2.10.1
+- Replaced the arg_parser code with gen_class.ArgParser code.
+
+### Changed
+- Multiple functions: Replaced the arg_parser code with gen_class.ArgParser code.
+- main: Removed gen_libs.get_inst call.
+
+
 ## [3.1.2] - 2022-11-22
 - Updated to work in Python 3 too
 - Upgraded python-lib to v2.9.4